move JCR-based eventing classes/interfaces into the modeshape impl
Resolves: https://jira.duraspace.org/browse/FCREPO-1866
