+++ fcrepo-kernel-modeshape/src/main/java/org/fcrepo/kernel/modeshape/observer/eventmappings/AllNodeEventsOneEvent.java
@@ -33,7 +33,6 @@ import javax.jcr.observation.Event;
 
 import org.fcrepo.kernel.api.exception.RepositoryRuntimeException;
 import org.fcrepo.kernel.api.observer.FedoraEvent;
-import org.fcrepo.kernel.api.observer.eventmappings.InternalExternalEventMapper;
 import org.fcrepo.kernel.modeshape.observer.FedoraEventImpl;
 
 import org.slf4j.Logger;
